// src/openapi/model/api-bank-transfer-application.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ApiBankTransferApplication
 */
export interface ApiBankTransferApplication {
    /**
     * 
     * @type {number}
     * @memberof ApiBankTransferApplication
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ApiBankTransferApplication
     */
    createdAt?: string;
    /**
     *
     * @type {number}
     * @memberof ApiBankTransferApplication
     */
    status: number;
    /**
     *
     * @type {number}
     * @memberof ApiBankTransferApplication
     */
    contractorId: number;
    /**
     *
     * @type {string}
     * @memberof ApiBankTransferApplication
     */
    contractorName: string;
    /**
     *
     * @type {string}
     * @memberof ApiBankTransferApplication
     */
    gmoBankDepositId: string;
    /**
     *
     * @type {string}
     * @memberof ApiBankTransferApplication
     */
    gmoBankId: string;
    /**
     *
     * @type {number}
     * @memberof ApiBankTransferApplication
     */
    gmoAmount: number;
    /**
     * 
     * @type {string}
     * @memberof ApiBankTransferApplication
     */
    gmoResultDetail: string;
    /**
     * 
     * @type {string}
     * @memberof ApiBankTransferApplication
     */
    transferExpectedDate: string;
    /**
     * 
     * @type {string}
     * @memberof ApiBankTransferApplication
     */
    transferCompleteDate?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiBankTransferApplication
     */
    gmoBankCode: string;
    /**
     * 
     * @type {string}
     * @memberof ApiBankTransferApplication
     */
    gmoBankName: string;
    /**
     * 
     * @type {string}
     * @memberof ApiBankTransferApplication
     */
    gmoBankBranchCode: string;
    /**
     * 
     * @type {string}
     * @memberof ApiBankTransferApplication
     */
    gmoBankBranchName: string;
    /**
     * 
     * @type {string}
     * @memberof ApiBankTransferApplication
     */
    gmoBankAccountType: string;
    /**
     * 
     * @type {string}
     * @memberof ApiBankTransferApplication
     */
    gmoBankAccountNumber: string;
    /**
     * 
     * @type {string}
     * @memberof ApiBankTransferApplication
     */
    gmoBankAccountName: string;
}