// src/openapi/model/api-mypage-balance.ts
/* tslint:disable */
/* eslint-disable */

/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 *
 * @export
 * @interface ApiMypageBalance
 */
export interface ApiMypageBalance {
    /**
     *
     * @type {number}
     * @memberof ApiMypageBalance
     */
    userId?: number;
    /**
     *
     * @type {string}
     * @memberof ApiMypageBalance
     */
    withDrawMonth?: string;
    /**
     *
     * @type {string}
     * @memberof ApiMypageBalance
     */
    transactionMonth?: string;
    /**
     *
     * @type {string}
     * @memberof ApiMypageBalance
     */
    depositDate?: string;
}