// src/openapi/model/api-payment.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ApiPayment
 */
export interface ApiPayment {
    /**
     * 
     * @type {number}
     * @memberof ApiPayment
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ApiPayment
     */
    createdAt?: string;
    /**
     * 
     * @type {number}
     * @memberof ApiPayment
     */
    ordererId: number;
    /**
     * 
     * @type {number}
     * @memberof ApiPayment
     */
    contractorId: number;
    /**
     * 
     * @type {number}
     * @memberof ApiPayment
     */
    projectApplicationId: number;
    /**
     * 
     * @type {number}
     * @memberof ApiPayment
     */
    projectId: number;
    /**
     * 
     * @type {number}
     * @memberof ApiPayment
     */
    paymentType: number;
    /**
     * 
     * @type {number}
     * @memberof ApiPayment
     */
    price: number;
    /**
     * 
     * @type {number}
     * @memberof ApiPayment
     */
    fee: number;
    /**
     * 
     * @type {number}
     * @memberof ApiPayment
     */
    contractorPay: number;
    /**
     * 
     * @type {number}
     * @memberof ApiPayment
     */
    status: number;
    /**
     * 
     * @type {string}
     * @memberof ApiPayment
     */
    gmoMemberId?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiPayment
     */
    gmoOrderId?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiPayment
     */
    gmoTranId?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiPayment
     */
    gmoTranDate?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiPayment
     */
    gmoCheckString?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiPayment
     */
    gmoErrCode?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiPayment
     */
    gmoErrInfo?: string;
}